refactor: split studio sections into components
